Minor quality of life changes in the clock handling code while it's playing life.
It used to display the clock, then run out the oscillation is had
already started and die 2 frames later, showing the clock again.  It
now lets the display keep going until the next time it would display
the clock, and checks a bool to see if it should display the clock or
break out of the loop.